chatterbot/requirements.txt
chatterbot==1.0.5
chatterbot-corpus==1.2.0
PyYAML==6.0
spacy==3.5.2
en-core-web-sm==3.5.0
requests==2.28.2
beautifulsoup4==4.11.2

